+++ redwood-tooling-codegen/src/main/kotlin/app/cash/redwood/tooling/codegen/composeGeneration.kt
@@ -30,6 +30,7 @@ import com.squareup.kotlinpoet.FileSpec
 import com.squareup.kotlinpoet.FunSpec
 import com.squareup.kotlinpoet.KModifier.INTERNAL
 import com.squareup.kotlinpoet.KModifier.OVERRIDE
+import com.squareup.kotlinpoet.KModifier.PUBLIC
 import com.squareup.kotlinpoet.ParameterSpec
 import com.squareup.kotlinpoet.ParameterizedTypeName.Companion.parameterizedBy
 import com.squareup.kotlinpoet.PropertySpec
@@ -71,6 +72,7 @@ internal fun generateComposable(
     addAnnotation(suppressDeprecations)
     addFunction(
       FunSpec.builder(flatName)
+        .addModifiers(if (widget.internalComposable) INTERNAL else PUBLIC)
         .addAnnotation(ComposeRuntime.Composable)
         .optIn(Redwood.RedwoodCodegenApi)
         .maybeAddKDoc(widget.documentation)