+++ redwood-tooling-schema/src/main/kotlin/app/cash/redwood/tooling/schema/schemaParserFir.kt
@@ -560,6 +560,7 @@ private fun FirContext.parseWidget(
     documentation = documentation,
     deprecation = deprecation,
     traits = traits,
+    internalComposable = annotation.internalComposable,
   )
 }
 
@@ -731,13 +732,22 @@ private fun FirContext.findWidgetAnnotation(
       (it as FirLiteralExpression).valueAsInt()
     }
 
-  return WidgetAnnotation(tagExpression.valueAsInt(), reservedProperties, reservedChildren)
+  val internalComposableExpression = annotation.argumentMapping
+    .mapping[Name.identifier("internalComposable")] as FirLiteralExpression?
+
+  return WidgetAnnotation(
+    tag = tagExpression.valueAsInt(),
+    reservedProperties = reservedProperties,
+    reservedChildren = reservedChildren,
+    internalComposable = internalComposableExpression?.valueAsBoolean() == true,
+  )
 }
 
 private data class WidgetAnnotation(
   val tag: Int,
   val reservedProperties: List<Int>,
   val reservedChildren: List<Int>,
+  val internalComposable: Boolean,
 )
 
 private fun FirContext.findPropertyAnnotation(
@@ -885,6 +895,7 @@ private fun ConeTypeProjection.toFqType(): FqType {
 }
 
 private fun FirLiteralExpression.valueAsInt() = (value as Long).toInt()
+private fun FirLiteralExpression.valueAsBoolean() = value as Boolean
 
 private fun ConeClassLikeType.extractParameterName(firSession: FirSession): String? {
   val annotation = customAnnotations